removed locking of sessions in database. Was causing problem on GAE and not working anyway. Eventually needs to be fixed.
git-svn-id: http://web2py.googlecode.com/svn/trunk@237 8fb14a07-5544-0410-8dbc-e3a1454e9f17

+++ gluon/contrib/gql.py
@@ -434,6 +434,7 @@ class SQLSet(object):
     def delete(self):
         if isinstance(self.sql_w,QueryException):
             item,fields=self._getitem_exception()
+            if not item: return
             item.delete()
         else:
             query,tablename,fields=self._select()
@@ -443,6 +444,7 @@ class SQLSet(object):
     def update(self,**update_fields):
         if isinstance(self.sql_w,QueryException):
             item,fields=self._getitem_exception()
+            if not item: return
             for key,value in update_fields.items():
                 setattr(item,key,value)
             item.put()
